+++ creme/decomposition/__init__.py
@@ -1,0 +1,9 @@
+"""
+Online matrix decomposition algorithms.
+"""
+from .lda import LDA
+
+
+__all__ = [
+    'LDA',
+]