+++ creme/metrics/__init__.py
@@ -5,20 +5,33 @@ loss in precision.
 from .accuracy import Accuracy
 from .accuracy import RollingAccuracy
 from .confusion import ConfusionMatrix
+from .confusion import RollingConfusionMatrix
 from .cross_entropy import CrossEntropy
 from .cross_entropy import RollingCrossEntropy
-from .f1_score import F1Score
-from .f1_score import MacroF1Score
-from .f1_score import MicroF1Score
-from .f1_score import RollingF1Score
-from .f1_score import RollingMacroF1Score
-from .f1_score import RollingMicroF1Score
+from .fbeta import F1
+from .fbeta import FBeta
+from .fbeta import MacroF1
+from .fbeta import MacroFBeta
+from .fbeta import MicroF1
+from .fbeta import MicroFBeta
+from .fbeta import MultiFBeta
+from .fbeta import RollingF1
+from .fbeta import RollingFBeta
+from .fbeta import RollingMacroF1
+from .fbeta import RollingMacroFBeta
+from .fbeta import RollingMicroF1
+from .fbeta import RollingMicroFBeta
+from .fbeta import RollingMultiFBeta
+from .jaccard import Jaccard
 from .log_loss import LogLoss
 from .log_loss import RollingLogLoss
 from .mae import MAE
 from .mae import RollingMAE
+from .mcc import MCC
+from .mcc import RollingMCC
 from .mse import MSE
 from .mse import RollingMSE
+from .multioutput import RegressionMultiOutput
 from .precision import MacroPrecision
 from .precision import MicroPrecision
 from .precision import Precision
@@ -35,6 +48,7 @@ from .rmse import RMSE
 from .rmse import RollingRMSE
 from .rmsle import RMSLE
 from .rmsle import RollingRMSLE
+from .roc_auc import ROCAUC
 from .smape import RollingSMAPE
 from .smape import SMAPE
 
@@ -43,32 +57,46 @@ __all__ = [
     'Accuracy',
     'ConfusionMatrix',
     'CrossEntropy',
-    'F1Score',
+    'F1',
+    'FBeta',
+    'Jaccard',
     'LogLoss',
-    'MacroF1Score',
+    'MAE',
+    'MacroF1',
+    'MacroFBeta',
     'MacroPrecision',
     'MacroRecall',
-    'MicroF1Score',
+    'MCC',
+    'MicroF1',
+    'MicroFBeta',
     'MicroPrecision',
     'MicroRecall',
-    'MAE',
     'MSE',
+    'MultiFBeta',
     'Precision',
     'Recall',
+    'RegressionMultiOutput',
     'RMSE',
     'RMSLE',
+    'ROCAUC',
     'RollingAccuracy',
+    'RollingConfusionMatrix',
     'RollingCrossEntropy',
-    'RollingF1Score',
+    'RollingF1',
+    'RollingFBeta',
     'RollingLogLoss',
-    'RollingMacroF1Score',
+    'RollingMAE',
+    'RollingMacroF1',
+    'RollingMacroFBeta',
     'RollingMacroPrecision',
     'RollingMacroRecall',
-    'RollingMAE',
-    'RollingMicroF1Score',
+    'RollingMCC',
+    'RollingMicroF1',
+    'RollingMicroFBeta',
     'RollingMicroPrecision',
     'RollingMicroRecall',
     'RollingMSE',
+    'RollingMultiFBeta',
     'RollingPrecision',
     'RollingRecall',
     'RollingRMSE',