+++ creme/metrics/mse.py
@@ -33,16 +33,15 @@ class MSE(stats.Mean, base.RegressionMetric):
 
     """
 
-    @property
-    def bigger_is_better(self):
-        return False
-
     def update(self, y_true, y_pred):
         return super().update((y_true - y_pred) ** 2)
 
 
 class RollingMSE(stats.RollingMean, base.RegressionMetric):
     """Rolling mean squared error.
+
+    Parameters:
+        window_size (int): Size of the window of recent values to consider.
 
     Example:
 
@@ -64,10 +63,6 @@ class RollingMSE(stats.RollingMean, base.RegressionMetric):
             0.5
 
     """
-
-    @property
-    def bigger_is_better(self):
-        return False
 
     def update(self, y_true, y_pred):
         return super().update((y_true - y_pred) ** 2)