+++ creme/multioutput/__init__.py
@@ -1,0 +1,11 @@
+"""
+Meta-estimators for multi-output classification and regression.
+"""
+from .chain import ClassifierChain
+from .chain import RegressorChain
+
+
+__all__ = [
+    'ClassifierChain',
+    'RegressorChain'
+]