+++ creme/stream.py
@@ -5,6 +5,12 @@ import csv
 import datetime as dt
 import itertools
 
+import numpy as np
+try:
+    import pandas as pd
+    PANDAS_INSTALLED = True
+except ImportError:
+    PANDAS_INSTALLED = False
 from sklearn import utils
 
 
@@ -17,13 +23,14 @@ __all__ = [
 ]
 
 
-def iter_numpy(X, y=None, feature_names=None, shuffle=False, random_state=None):
+def iter_numpy(X, y=None, feature_names=None, target_names=None, shuffle=False, random_state=None):
     """Yields rows from an array of features and an array of targets.
 
     Parameters:
         X (array-like of shape (n_samples, n_features))
         y (array-like of shape (n_samples,))
         feature_names (list of length n_features)
+        target_names (list of length n_outputs)
         shuffle (bool): Whether to shuffle the inputs or not.
         random_state (int, RandomState instance or None, default=None): If int, ``random_state`` is
             the seed used by the random number generator; if ``RandomState`` instance,
@@ -35,31 +42,44 @@ def iter_numpy(X, y=None, feature_names=None, shuffle=False, random_state=None):
 
     """
     feature_names = list(range(len(X[0]))) if feature_names is None else feature_names
-    rng = utils.check_random_state(random_state)
 
+    multioutput = y is not None and not np.isscalar(y[0])
+    if multioutput and target_names is None:
+        target_names = list(range(len(y[0])))
+
+    # Shuffle the data
+    rng = utils.check_random_state(random_state)
     if shuffle:
         order = rng.permutation(len(X))
         X, y = X[order], y if y is None else y[order]
 
-    for x, yi in itertools.zip_longest(X, y if hasattr(y, '__iter__') else []):
-        yield dict(zip(feature_names, x)), yi
+    if multioutput:
+
+        for x, yi in itertools.zip_longest(X, y if hasattr(y, '__iter__') else []):
+            yield dict(zip(feature_names, x)), dict(zip(target_names, yi))
+
+    else:
+
+        for x, yi in itertools.zip_longest(X, y if hasattr(y, '__iter__') else []):
+            yield dict(zip(feature_names, x)), yi
 
 
-def iter_sklearn_dataset(load_dataset, **kwargs):
+def iter_sklearn_dataset(dataset, **kwargs):
     """Yields rows from one of the datasets provided by scikit-learn.
 
     Parameters:
-        load_dataset (callable): The method used to load the dataset, for example
-        `sklearn.datasets.load_boston`.
+        dataset (sklearn.utils.Bunch): A scikit-learn dataset.
 
     Yields:
         tuple: A pair (``x``, ``y``) where ``x`` is a dict of features and ``y`` is the target.
 
     """
-    dataset = load_dataset()
     kwargs['X'] = dataset.data
     kwargs['y'] = dataset.target
-    kwargs['feature_names'] = dataset.feature_names
+    try:
+        kwargs['feature_names'] = dataset.feature_names
+    except AttributeError:
+        pass
 
     for x, yi in iter_numpy(**kwargs):
         yield x, yi
@@ -77,6 +97,9 @@ def iter_pandas(X, y=None, **kwargs):
 
     """
     kwargs['feature_names'] = X.columns
+    if isinstance(y, pd.DataFrame):
+        kwargs['target_names'] = y.columns
+
     for x, yi in iter_numpy(X.to_numpy(), y, **kwargs):
         yield x, yi